Add 64 bit memory support

+++ crates/wast/src/encode.rs
@@ -38,8 +38,8 @@ impl Encode for str {
 
 impl Encode for usize {
     fn encode(&self, e: &mut Vec<u8>) {
-        assert!(*self <= u32::max_value() as usize);
-        (*self as u32).encode(e)
+        assert!(*self <= u64::max_value() as usize);
+        (*self as u64).encode(e)
     }
 }
